requirements.txt
langchain==0.0.205
google-api-python-client==2.89.0
openai==0.27.8
python-dotenv==1.0.0